+++ tests/test_prepare.py
@@ -15,6 +15,7 @@
 """NEMO-Cmd prepare sub-command plug-in unit tests
 """
 import os
+from pathlib import Path
 try:
     from unittest.mock import call, Mock, patch
 except ImportError:
@@ -707,49 +708,139 @@ class TestMakeExecutableLinks:
 
 
 class TestMakeGridLinks:
-    @patch('nemo_cmd.prepare.logger')
-    def test_no_forcing_dir(self, m_log):
-        run_desc = {'paths': {'forcing': 'foo'}}
-        nemo_cmd.prepare._remove_run_dir = Mock()
-        p_exists = patch('nemo_cmd.prepare.os.path.exists', return_value=False)
-        p_abspath = patch(
-            'nemo_cmd.prepare.os.path.abspath', side_effect=lambda path: path
-        )
-        with pytest.raises(SystemExit), p_exists, p_abspath:
-            nemo_cmd.prepare._make_grid_links(run_desc, 'run_dir')
-        m_log.error.assert_called_once_with(
-            'foo not found; cannot create symlinks - '
-            'please check the forcing path in your run description file'
-        )
-        nemo_cmd.prepare._remove_run_dir.assert_called_once_with('run_dir')
+    """Unit tests for `nemo prepare` _make_grid_links() function.
+    """
 
+    @patch('nemo_cmd.prepare._remove_run_dir')
     @patch('nemo_cmd.prepare.logger')
-    def test_no_link_path(self, m_log):
+    def test_no_grid_coordinates_key(self, m_logger, m_rm_run_dir):
+        run_desc = {}
+        with pytest.raises(SystemExit):
+            nemo_cmd.prepare._make_grid_links(run_desc, 'run_dir')
+        assert m_logger.error.call_args_list[0] == call(
+            'grid: coordinates key not found - '
+            'please check your run description YAML file'
+        )
+        m_rm_run_dir.assert_called_once_with('run_dir')
+
+    @patch('nemo_cmd.prepare._remove_run_dir')
+    @patch('nemo_cmd.prepare.logger')
+    def test_no_grid_bathymetry_key(self, m_logger, m_rm_run_dir):
+        run_desc = {'grid': {'coordinates': 'coords.nc'}}
+        with pytest.raises(SystemExit):
+            nemo_cmd.prepare._make_grid_links(run_desc, 'run_dir')
+        assert m_logger.error.call_args_list[0] == call(
+            'grid: bathymetry key not found - '
+            'please check your run description YAML file'
+        )
+        m_rm_run_dir.assert_called_once_with('run_dir')
+
+    @patch('nemo_cmd.prepare._remove_run_dir')
+    @patch('nemo_cmd.prepare.logger')
+    def test_no_forcing_key(self, m_logger, m_rm_run_dir):
+        run_desc = {
+            'grid': {
+                'coordinates': 'coords.nc',
+                'bathymetry': 'bathy.nc'
+            }
+        }
+        with pytest.raises(SystemExit):
+            nemo_cmd.prepare._make_grid_links(run_desc, 'run_dir')
+        m_logger.error.assert_called_once_with(
+            'forcing key not found - '
+            'please check your run description YAML file'
+        )
+        m_rm_run_dir.assert_called_once_with('run_dir')
+
+    @patch('nemo_cmd.prepare._remove_run_dir')
+    @patch('nemo_cmd.prepare.logger')
+    def test_no_forcing_dir(self, m_logger, m_rm_run_dir):
         run_desc = {
             'paths': {
-                'forcing': 'foo',
+                'forcing': '/foo'
             },
             'grid': {
-                'coordinates': 'coordinates.nc',
-                'bathymetry': 'bathy.nc',
+                'coordinates': 'coords.nc',
+                'bathymetry': 'bathy.nc'
+            }
+        }
+        p_exists = patch('nemo_cmd.prepare.Path.exists', return_value=False)
+        with pytest.raises(SystemExit), p_exists:
+            nemo_cmd.prepare._make_grid_links(run_desc, 'run_dir')
+        m_logger.error.assert_called_once_with(
+            '/foo not found; cannot create symlinks - '
+            'please check the forcing path in your run description file'
+        )
+        m_rm_run_dir.assert_called_once_with('run_dir')
+
+    @patch('nemo_cmd.prepare._remove_run_dir')
+    @patch('nemo_cmd.prepare.logger')
+    def test_no_link_path_absolute_coords_bathy(self, m_logger, m_rm_run_dir):
+        run_desc = {
+            'grid': {
+                'coordinates': '/coords.nc',
+                'bathymetry': '/bathy.nc'
             },
         }
-        nemo_cmd.prepare._remove_run_dir = Mock()
-        p_exists = patch(
-            'nemo_cmd.prepare.os.path.exists', side_effect=[True, False]
-        )
-        p_abspath = patch(
-            'nemo_cmd.prepare.os.path.abspath', side_effect=lambda path: path
-        )
-        p_chdir = patch('nemo_cmd.prepare.os.chdir')
-        with pytest.raises(SystemExit), p_exists, p_abspath, p_chdir:
+        with pytest.raises(SystemExit):
             nemo_cmd.prepare._make_grid_links(run_desc, 'run_dir')
-        m_log.error.assert_called_once_with(
-            'foo/grid/coordinates.nc not found; cannot create symlink - '
+        m_logger.error.assert_called_once_with(
+            '/coords.nc not found; cannot create symlink - '
             'please check the forcing path and grid file names '
             'in your run description file'
         )
-        nemo_cmd.prepare._remove_run_dir.assert_called_once_with('run_dir')
+        m_rm_run_dir.assert_called_once_with('run_dir')
+
+    @patch('nemo_cmd.prepare._remove_run_dir')
+    @patch('nemo_cmd.prepare.logger')
+    def test_no_link_path_relative_coords_bathy(
+        self, m_logger, m_rm_run_dir, tmpdir
+    ):
+        forcing_dir = tmpdir.ensure_dir('foo')
+        grid_dir = forcing_dir.ensure_dir('grid')
+        run_dir = tmpdir.ensure_dir('runs')
+        run_desc = {
+            'paths': {
+                'forcing': str(forcing_dir)
+            },
+            'grid': {
+                'coordinates': 'coords.nc',
+                'bathymetry': 'bathy.nc'
+            },
+        }
+        with pytest.raises(SystemExit):
+            nemo_cmd.prepare._make_grid_links(run_desc, str(run_dir))
+        m_logger.error.assert_called_once_with(
+            '{}/coords.nc not found; cannot create symlink - '
+            'please check the forcing path and grid file names '
+            'in your run description file'.format(grid_dir)
+        )
+        m_rm_run_dir.assert_called_once_with(run_dir)
+
+    @patch('nemo_cmd.prepare._remove_run_dir')
+    @patch('nemo_cmd.prepare.logger')
+    def test_link_path(self, m_logger, m_rm_run_dir, tmpdir):
+        forcing_dir = tmpdir.ensure_dir('foo')
+        grid_dir = forcing_dir.ensure_dir('grid')
+        grid_dir.ensure('coords.nc')
+        grid_dir.ensure('bathy.nc')
+        run_dir = tmpdir.ensure_dir('runs')
+        run_desc = {
+            'paths': {
+                'forcing': str(forcing_dir)
+            },
+            'grid': {
+                'coordinates': 'coords.nc',
+                'bathymetry': 'bathy.nc'
+            },
+        }
+        nemo_cmd.prepare._make_grid_links(run_desc, str(run_dir))
+        assert Path(str(run_dir), 'coordinates.nc').is_symlink()
+        assert Path(str(run_dir), 'coordinates.nc'
+                    ).samefile(str(grid_dir.join('coords.nc')))
+        assert Path(str(run_dir), 'bathy_meter.nc').is_symlink()
+        assert Path(str(run_dir), 'bathy_meter.nc'
+                    ).samefile(str(grid_dir.join('bathy.nc')))
 
 
 class TestMakeForcingLinks: